requirements.txt
playsound >=1.3.0
pyusb >=1.3.1
numpy >=2.3.0
paho-mqtt >= 2.1.0
notify-py >= 0.3.43